version 1.0.2

+++ lib/data_layer/data_source/remote/order_service.dart
@@ -43,7 +43,7 @@ class OrderService extends BaseService {
   /// 提现  收益 申请提现
   AsyncJson incomeApplyWithdraw({
     required int cardId,
-    required int amount,
+    required double amount,
     required int type,
   }) =>
       post('/withdraw', data: {
